Struct subotai::hash::SubotaiHash

pub struct SubotaiHash {
    pub raw: [u8; 20],
}

Subotai hash, a light wrapper over a little endian HASH_SIZE bit hash. It can be generated randomly or via sha-1 of a given string.

Fields
raw: [u8; 20]
Methods

impl SubotaiHash

fn blank() -> SubotaiHash

Generates a blank hash (every bit set to 0).

fn random() -> SubotaiHash

Generates a random hash via kernel supplied entropy.

fn sha1(data: &str) -> SubotaiHash

Generates a SHA-1 hash from a string.

fn random_at_distance(reference: &SubotaiHash, distance: usize) -> SubotaiHash

Creates a random hash at a given XOR distance from another (height of their XOR value).

fn zeroes(&self) -> Zeroes

Provides an iterator through the indices of each of its "0" bits.

fn ones(&self) -> Ones

Provides an iterator through the indices of each of its "1" bits.

fn into_zeroes(self) -> IntoZeroes

Provides a consuming iterator through the indices of each of its "0" bits.

fn into_ones(self) -> IntoOnes

Provides a consuming iterator through the indices of each of its "1" bits.

fn height(&self) -> Option<usize>

Computes the bit index of the highest "1". Returns None for a blank hash.

fn flip_bit(&mut self, position: usize)

Flips a bit in the hash.