Module subotai::node

The node module is the main point of contact with Subotai. Node structs contain all the pieces needed to join a Subotai network.

When you initialize a Node struct, a few threads are automatically launched that take care of listening to RPCs from other nodes, automatic maintenance and eviction of older node entries, and more.

Nodes start in the OffGrid state by default, meaning they aren't associated to a network. You must bootstrap the node by providing the address of another node, at which point the state will change to OnGrid.

Destroying a node automatically schedules all threads to terminate after finishing any pending operations.

Modules

receptions
Allows listening to RPCs received by a node. Unnecessary for normal operation, but it can be useful for debugging your network.

Structs

Configuration
Network configuration constants. Do not set these values directly, as there is no way to initialize a node from a Configuration struct. Instead, use node::Factory if you want your application to use non-default network constants.

Factory
Allows the construction of nodes with custom network constants, specific ports, and other options.

Node
Subotai node.

NodeInfo
ID - Address pair that identifies a unique Subotai node in the network.

Enums

State
State of a Subotai node.

StorageEntry
This is the data type that can be stored and retrieved in the Subotai network, consisting of either another hash or a binary blob.

Constants

SOCKET_BUFFER_SIZE_BYTES
Size of a typical UDP socket buffer.