Struct subotai::node::Configuration

pub struct Configuration {
    pub alpha: usize,
    pub impatience: usize,
    pub k_factor: usize,
    pub max_conflicts: usize,
    pub max_storage: usize,
    pub max_storage_blob_size: usize,
    pub expiration_distance_threshold: usize,
    pub base_expiration_time_hrs: i64,
    pub base_cache_time_mins: i64,
    pub network_timeout_s: i64,
}

Network configuration constants. Do not set these values directly, as there is no way to initialize a node from a Configuration struct. Instead, use node::Factory if you want your application to use non-default network constants.

Note that for the network to function optimally, the alpha, impatience, expiration_distance_threshold and base_expiration_time_hrs must be identical for all nodes.

Fields
alpha: usize

Network-wide concurrency factor. It's used, for example, to decide the number of remote nodes to interrogate concurrently when performing a network-wide lookup.

impatience: usize

Impatience factor, valid in the range [0..ALPHA). When performing "waves", the impatience factor denotes how many nodes we can give up waiting for, before starting the next wave.

If we send a request to ALPHA nodes during a lookup wave, we will start the next wave after we receive 'ALPHA - IMPATIENCE' responses.

k_factor: usize

Data structure factor. It's used to dictate the size of the internal routing data structures (k-buckets).

max_conflicts: usize

Maximum amount of eviction conflicts allowed before the node goes into a temporary defensive mode, and starts to prioritize old contacts to new, potentially malicious ones.

max_storage: usize

Maximum amount of storage entries (key-value or key-blob pairs). This has no effect on the routing table size (amount of node id-address pairs), which is dictated by the k_factor.

max_storage_blob_size: usize

Maximum size in bytes for a blob storage entry. (A blob entry consists in a key associated with a chunk of binary data, instead of a 160 bit value hash).

expiration_distance_threshold: usize

Xor distance from a key at which point nodes will start to dramatically decrease the expiration time for cached storage entries. This is only relevant in cases of extreme network traffic around a given key. A bigger threshold allows for more over-caching and therefore less load on the critical nodes, while a smaller threshold allows for a leaner network.

base_expiration_time_hrs: i64

Base expiration time for storage entries. Every time you call store on a node that resides on a live network (i.e. is in an OnGrid state) you guarantee the entry will remain in the network for this number of hours. Calling store again will refresh this time.

base_cache_time_mins: i64

Base expiration time for cached storage entries. When several nodes attempt to retrieve the same entry, it is cached at progressively longer distances from the owner. This will not prolong the overall lifespan of the entry because cached entries do not live long enough to be republished.

network_timeout_s: i64

Time in seconds after which it can be assumed that a remote node has failed to respond to a query.