Struct subotai::node::Factory

pub struct Factory { /* fields omitted */ }

Allows the construction of nodes with custom network constants, specific ports, and other options.

Methods

impl Factory

fn new() -> Self

fn create_node(&self) -> SubotaiResult<Node>

Creates a node with the configuration values specified in the factory. Defaults to the same values as calling Node::new().

fn inbound_port(self, port: u16) -> Self

Inbound UDP port for incoming RPCs.

fn outbound_port(self, port: u16) -> Self

Outbound UDP port for outgoing RPCs.

fn alpha(self, alpha: usize) -> Self

Network-wide concurrency factor. It's used, for example, to decide the number of remote nodes to interrogate concurrently when performing a network-wide lookup.

fn impatience(self, impatience: usize) -> Self

Impatience factor, valid in the range [0..ALPHA). When performing "waves", the impatience factor denotes how many nodes we can give up waiting for, before starting the next wave.

If we send a request to ALPHA nodes during a lookup wave, we will start the next wave after we receive 'ALPHA - IMPATIENCE' responses.

fn k_factor(self, k_factor: usize) -> Self

Data structure factor. It's used to dictate the size of the internal routing data structures (k-buckets).

fn max_conflicts(self, max_conflicts: usize) -> Self

Maximum amount of eviction conflicts allowed before the node goes into a temporary defensive mode, and starts to prioritize old contacts to new, potentially malicious ones.

fn max_storage(self, max_storage: usize) -> Self

Maximum amount of storage entries (key-value or key-blob pairs). This has no effect on the routing table size (amount of node id-address pairs), which is dictated by the k_factor.

fn max_storage_blob_size(self, max_storage_blob_size: usize) -> Self

Maximum size in bytes for a blob storage entry. (A blob entry consists in a key associated with a chunk of binary data, instead of a 160 bit value hash).

fn expiration_distance_threshold(self, expiration_distance_threshold: usize) -> Self

Xor distance from a key at which point nodes will start to dramatically decrease the expiration time for cached storage entries. This is only relevant in cases of extreme network traffic around a given key. A bigger threshold allows for more over-caching and therefore less load on the critical nodes, while a smaller threshold allows for a leaner network.

fn network_timeout_s(self, network_timeout_s: i64) -> Self

Time in seconds after which it can be assumed that a remote node has failed to respond to a query.

fn base_expiration_time_hrs(self, base_expiration_time_hrs: i64) -> Self

Base expiration time for storage entries. Every time you call store on a node that resides on a live network (i.e. is in an OnGrid state) you guarantee the entry will remain in the network for this number of hours. Calling store again will refresh this time.

fn base_cache_time_mins(self, base_cache_time_mins: i64) -> Self

Base expiration time for cached storage entries. When several nodes attempt to retrieve the same entry, it is cached at progressively longer distances from the owner. This will not prolong the overall lifespan of the entry because cached entries do not live long enough to be republished.

Note: Values will be clamped to always be smaller than 1 hour.