Struct subotai::node::Node

pub struct Node { /* fields omitted */ }

Subotai node.

Methods

impl Node

fn new() -> SubotaiResult<Node>

Constructs a node with OS allocated random ports and default network constants.

If you need more control over ports and network configuration, use node::Factory.

fn store(&self, key: SubotaiHash, entry: StorageEntry) -> SubotaiResult<()>

Stores an entry in the network, refreshing its expiration time back to the base value.

fn retrieve(&self, key: &SubotaiHash) -> SubotaiResult<Vec<StorageEntry>>

Retrieves all values associated to a key from the network.

fn id(&self) -> &SubotaiHash

Returns the hash used to identify this node in the network.

fn configuration(&self) -> &Configuration

Returns the network constant strucure for this node.

fn state(&self) -> State

Returns the current state of the node.

fn receptions(&self) -> Receptions

Produces an iterator over RPCs received by this node. The iterator will block indefinitely.

fn bootstrap(&self, seed: &SocketAddr) -> SubotaiResult<()>

Bootstraps the node from a seed IP:Port pair. Returns Ok(()) if the seed has been reached and the asynchronous bootstrap process has started. However, it might take a bit for the node to become alive (use node::wait_until_state to block until it's alive, if necessary).

fn wait_for_state(&self, state: State)

Returns if the node is already in the specified state, otherwise blocks indefinitely until that state is reached.

fn local_info(&self) -> NodeInfo

Retrieves the node ID + address pair.